refactor: gbs line up page
